Revert "Enable per-spw gain solutions beyond the inf_EB stage"
